fix types import incorrectly updated by moving the files
