+++ web/src/routes/create_account/+page.svelte
@@ -1,13 +1,7 @@
 <script lang="ts">
-	import { enhance } from "$app/forms";
-	import { goto } from "$app/navigation";
-
-	import FormButton from "$lib/components/form/FormButton.svelte";
-	import Input from "$lib/components/form/Input.svelte";
 	import H1 from "$lib/components/text/H1.svelte";
-	import { toasts, newToast, ToastType } from "$lib/state/toast.svelte.js";
 	import SignupForm from "./SignupForm.svelte";
-	import type { PageData } from "../../../../.svelte-kit/types/src/routes";
+	import type { PageData } from "./$types.js";
 
 	let { data }: { data: PageData } = $props();
 </script>
