+++ web/src/routes/events/[id]/+page.server.ts
