+++ web/src/routes/sign_in/+page.svelte
@@ -2,7 +2,7 @@
 	import H1 from "$lib/components/text/H1.svelte";
 	import Link from "$lib/components/text/Link.svelte";
 
-	import type { PageData } from "../../../../.svelte-kit/types/src/routes";
+	import type { PageData } from "./$types.js";
 	import SigninForm from "./SigninForm.svelte";
 
 	let { data }: { data: PageData } = $props();
