+++ web/src/routes/users/edit/profile_picture/+page.server.ts
@@ -1,6 +1,6 @@
 import { getUser } from "$lib/api_calls/users.svelte.js";
 import * as api from "$lib/api_calls/api.svelte.js";
-import type { PageServerLoad } from "../../../../../.svelte-kit/types/src/routes";
+import type { PageServerLoad } from "./$types.js";
 import { superValidate } from "sveltekit-superforms";
 import { imageFormSchema } from "./schema";
 import { zod } from "sveltekit-superforms/adapters";
